pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "santafe-da"
version = "2.0.0"
description = "Santa Fe Double Auction Tournament - 1993 Replication with Modern AI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Santa Fe DA Project" }
]
keywords = ["double-auction", "market-simulation", "reinforcement-learning", "agent-based-modeling"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core scientific computing
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "polars>=0.19.0",
    # Configuration management
    "hydra-core>=1.3.0",
    # Deep Learning & RL
    "torch>=2.0.0",
    "stable-baselines3>=2.0.0",
    "gymnasium>=0.29.0",
    # LLM Integration
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    # Experiment tracking & visualization
    "wandb>=0.15.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    # Utilities
    "tqdm>=4.65.0",
    "tabulate>=0.9.0",
    "sb3-contrib>=2.0.0",
    "pypdf>=6.3.0",
    "scipy>=1.15.3",
    # Visual testing
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",

    # Type checking
    "mypy>=1.5.0",
    "types-tabulate",

    # Linting & formatting
    "ruff>=0.1.0",
    "black>=23.7.0",

    # Pre-commit hooks
    "pre-commit>=3.3.0",
]

[project.urls]
Homepage = "https://github.com/santafe-da/santafe-1"
Repository = "https://github.com/santafe-da/santafe-1"

# Package discovery - use root-level modules
[tool.hatch.build.targets.wheel]
packages = ["engine", "traders", "envs"]

# ============================================================================
# Tool Configurations
# ============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Exclude directories that contain test-like files but aren't part of the test suite
norecursedirs = [
    ".git",
    ".venv",
    "venv",
    "oldcode",
    "configs",
    "tools",
    "experiments",
    "results",
    "__pycache__",
]
addopts = [
    "--verbose",
    "--tb=short",
    "--cov=engine",
    "--cov=traders",
    "--cov=envs",
    "--cov-report=html",
    "--cov-report=term-missing",
    #"--cov-fail-under=90",  # Disabled for Phase 0 - will enable in Phase 1
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow-running tests",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = false
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "stable_baselines3.*",
    "gymnasium.*",
    "wandb.*",
    "hydra.*",
    "litellm.*",
    "polars.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip without explicit strict parameter
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/*" = ["ARG", "SIM"]  # Relax rules for test files

[tool.ruff.isort]
known-first-party = ["engine", "traders", "envs"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.ruff_cache
  | \.venv
  | _build
  | build
  | dist
  | oldcode
)/
'''

[tool.coverage.run]
source = ["engine", "traders", "envs"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 90
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
